[feat] 유저 앱에 대해서도 같은 처리 수행
